simplify pass 1: unnest proto manager
