+++ packages/insomnia/src/ui/components/proto-file/proto-file-list.tsx
@@ -8,8 +8,8 @@ import { ProtoDirectoryListItem } from './proto-directory-list-item';
 import { ProtoFileListItem } from './proto-file-list-item';
 
 export type SelectProtoFileHandler = (id: string) => void;
-export type DeleteProtoFileHandler = (protofile: ProtoFile) => Promise<void>;
-export type DeleteProtoDirectoryHandler = (protoDirectory: ProtoDirectory) => Promise<void>;
+export type DeleteProtoFileHandler = (protofile: ProtoFile) => void;
+export type DeleteProtoDirectoryHandler = (protoDirectory: ProtoDirectory) => void;
 export type UpdateProtoFileHandler = (protofile: ProtoFile) => Promise<void>;
 export type RenameProtoFileHandler = (protoFile: ProtoFile, name?: string) => Promise<void>;
 
@@ -18,7 +18,6 @@ interface Props {
   selectedId?: string;
   handleSelect: SelectProtoFileHandler;
   handleDelete: DeleteProtoFileHandler;
-  handleRename: RenameProtoFileHandler;
   handleUpdate: UpdateProtoFileHandler;
   handleDeleteDirectory: DeleteProtoDirectoryHandler;
 }
@@ -31,7 +30,6 @@ const recursiveRender = (
   const {
     handleDelete,
     handleDeleteDirectory,
-    handleRename,
     handleSelect,
     handleUpdate,
     selectedId,
@@ -51,7 +49,6 @@ const recursiveRender = (
       isSelected={f._id === selectedId}
       handleSelect={handleSelect}
       handleDelete={handleDelete}
-      handleRename={handleRename}
       handleUpdate={handleUpdate}
       indentLevel={indent}
     />
